changed Disc flight number properties from Int16 to Double

+++ DiscDrawer/Models/Disc+CoreData/Disc+CoreDataProperties.swift
@@ -17,16 +17,16 @@ extension Disc {
     }
 
     @NSManaged public var condition: String?
-    @NSManaged public var fade: Int16
+    @NSManaged public var fade: Double
     @NSManaged public var flightChartUrl: String?
-    @NSManaged public var glide: Int16
+    @NSManaged public var glide: Double
     @NSManaged public var inBag: Bool
     @NSManaged public var manufacturer: String?
     @NSManaged public var name: String?
     @NSManaged public var plastic: String?
-    @NSManaged public var speed: Int16
+    @NSManaged public var speed: Double
     @NSManaged public var stability: String?
-    @NSManaged public var turn: Int16
+    @NSManaged public var turn: Double
     @NSManaged public var type: String?
     @NSManaged public var weight: Int16
